\name{EnKFP-package}
\alias{EnKFP-package}
\alias{EnKFP}
\docType{package}
\title{
\packageTitle{EnKFP}
}
\description{
\packageDescription{EnKFP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EnKFP}
\packageIndices{EnKFP}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EnKFP}

Maintainer: \packageMaintainer{EnKFP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
